breaking intervals and estimation up into test and code files
